add : full service configuration

+++ price-producer/internal/domain/dto/price.go
@@ -2,5 +2,5 @@ package dto
 
 type AddPriceRequest struct {
 	Symbol string  `json:"symbol"`
-	Value  float32 `json:"value"`
+	Value  float64 `json:"value"`
 }
